+++ CHANGELOG.md
@@ -39,9 +39,8 @@ develop
     - Show / Hide Status Bar
 - Remove feature that changes the line height of current document from the “Format” menu.
     - From this, `line spacing` property on AppleScript is also deprecated.
-- Remove  “Set as Default” button in the editor opacity panel.
-- Update preferences icons.
-- Use small size toolbar icons as default.
+- Remove “Set as Default” button in the editor opacity panel.
+- Update preferences icons and toolbar.
 - Update key binding setting format.
     - Not compatible with previous key bindings setting. Please customize again in the preferences window.
 - Change specification not to treat full-width spaces as indent.